+++ client/cli/src/bounty.rs
@@ -1,26 +1,33 @@
-use crate::error::{Error, Result};
+use crate::error::{
+    Error,
+    Result,
+};
 use clap::Clap;
 use core::fmt::{
     Debug,
     Display,
 };
-use libipld::{
-    cid::{
-        Cid,
-        Codec,
-    },
-    multihash::Blake2b256,
-};
 use substrate_subxt::{
     sp_core::crypto::Ss58Codec,
     system::System,
+    Runtime,
 };
-use util::court::ResolutionMetadata;
-use utils_identity::cid::CidBytes;
+use sunshine_bounty_client::{
+    bank::Bank,
+    bounty::{
+        Bounty,
+        BountyClient,
+    },
+    org::Org,
+    vote::Vote,
+};
+use sunshine_bounty_utils::court::ResolutionMetadata;
 
 #[derive(Clone, Debug, Clap)]
 pub struct BountyPostCommand {
-    pub description: String,
+    pub repo_owner: String,
+    pub repo_name: String,
+    pub issue_number: u64,
     pub amount_reserved_for_bounty: u128,
     // ac == acceptance committee
     pub ac_org: u64,
@@ -34,41 +41,35 @@ pub struct BountyPostCommand {
     pub sc_duration: Option<u32>,
 }
 
-#[async_trait]
-impl<T: Runtime + Org + Vote + Donate + Bank + Bounty, P: Pair> Command<T, P>
-    for BountyPostCommand
-where
-    <T as System>::AccountId: Ss58Codec,
-    <T as System>::BlockNumber: From<u32> + Display,
-    <T as Vote>::Signal: From<u64> + Display,
-    <T as Org>::OrgId: From<u64> + Display,
-    <T as Org>::IpfsReference: From<CidBytes> + Debug,
-    <T as Bank>::Currency: From<u128> + Display,
-    <T as Bounty>::BountyId: Display,
-{
-    async fn exec(&self, client: &dyn AbstractClient<T, P>) -> Result<()> {
-        let description: CidBytes = {
-            let content = self.description.as_bytes();
-            let hash = Blake2b256::digest(&content[..]);
-            let cid = Cid::new_v1(Codec::Raw, hash);
-            CidBytes::from(&cid)
-        };
-        let ac_rejection_threshold: Option<T::Signal> =
+impl BountyPostCommand {
+    pub async fn exec<R: Runtime + Bounty, C: BountyClient<R>>(
+        &self,
+        client: &C,
+    ) -> Result<(), C::Error>
+    where
+        <R as System>::AccountId: Ss58Codec,
+        <R as System>::BlockNumber: From<u32> + Display,
+        <R as Vote>::Signal: From<u64> + Display,
+        <R as Org>::OrgId: From<u64> + Display,
+        <R as Bank>::Currency: From<u128> + Display,
+        <R as Bounty>::BountyId: Display,
+    {
+        let ac_rejection_threshold: Option<R::Signal> =
             if let Some(ac_r_t) = self.ac_rejection_threshold {
                 Some(ac_r_t.into())
             } else {
                 None
             };
-        let ac_duration: Option<T::BlockNumber> =
+        let ac_duration: Option<R::BlockNumber> =
             if let Some(ac_d) = self.ac_duration {
                 Some(ac_d.into())
             } else {
                 None
             };
         let acceptance_committee: ResolutionMetadata<
-            <T as Org>::OrgId,
-            <T as Vote>::Signal,
-            <T as System>::BlockNumber,
+            <R as Org>::OrgId,
+            <R as Vote>::Signal,
+            <R as System>::BlockNumber,
         > = ResolutionMetadata::new(
             self.ac_org.into(),
             self.ac_passage_threshold.into(),
@@ -77,21 +78,21 @@ where
         );
         let supervision_committee: Option<
             ResolutionMetadata<
-                <T as Org>::OrgId,
-                <T as Vote>::Signal,
-                <T as System>::BlockNumber,
+                <R as Org>::OrgId,
+                <R as Vote>::Signal,
+                <R as System>::BlockNumber,
             >,
         > = if let Some(org) = self.sc_org {
             let passage_threshold = self
                 .sc_passage_threshold
                 .ok_or(Error::PostBountyInputError)?;
-            let sc_rejection_threshold: Option<T::Signal> =
+            let sc_rejection_threshold: Option<R::Signal> =
                 if let Some(sc_r_t) = self.sc_rejection_threshold {
                     Some(sc_r_t.into())
                 } else {
                     None
                 };
-            let sc_duration: Option<T::BlockNumber> =
+            let sc_duration: Option<R::BlockNumber> =
                 if let Some(sc_d) = self.sc_duration {
                     Some(sc_d.into())
                 } else {
@@ -108,12 +109,15 @@ where
         };
         let event = client
             .account_posts_bounty(
-                description.into(),
+                (*self.repo_owner).to_string(),
+                (*self.repo_name).to_string(),
+                self.issue_number,
                 self.amount_reserved_for_bounty.into(),
                 acceptance_committee,
                 supervision_committee,
             )
-            .await?;
+            .await
+            .map_err(Error::Client)?;
         println!(
             "AccountId {} posted new bounty with identifier {} with amount reserved: {}",
             event.poster, event.new_bounty_id, event.amount_reserved_for_bounty
@@ -129,29 +133,24 @@ pub struct BountyApplicationCommand {
     pub total_amount: u128,
 }
 
-#[async_trait]
-impl<T: Runtime + Org + Vote + Donate + Bank + Bounty, P: Pair> Command<T, P>
-    for BountyApplicationCommand
-where
-    <T as System>::AccountId: Ss58Codec,
-    <T as Org>::IpfsReference: From<CidBytes> + Debug,
-    <T as Bank>::Currency: From<u128> + Display,
-    <T as Bounty>::BountyId: From<u64> + Display,
-{
-    async fn exec(&self, client: &dyn AbstractClient<T, P>) -> Result<()> {
-        let description: CidBytes = {
-            let content = self.description.as_bytes();
-            let hash = Blake2b256::digest(&content[..]);
-            let cid = Cid::new_v1(Codec::Raw, hash);
-            CidBytes::from(&cid)
-        };
+impl BountyApplicationCommand {
+    pub async fn exec<R: Runtime + Bounty, C: BountyClient<R>>(
+        &self,
+        client: &C,
+    ) -> Result<(), C::Error>
+    where
+        <R as System>::AccountId: Ss58Codec,
+        <R as Bank>::Currency: From<u128> + Display,
+        <R as Bounty>::BountyId: From<u64> + Display,
+    {
         let event = client
             .account_applies_for_bounty(
                 self.bounty_id.into(),
-                description.into(),
+                (*self.description).to_string(),
                 self.total_amount.into(),
             )
-            .await?;
+            .await
+            .map_err(Error::Client)?;
         println!(
             "AccountId {:?} applied for bounty with identifier {} with application identifier {} for total amount {}",
             event.submitter, event.bounty_id, event.new_grant_app_id, event.total_amount,
@@ -166,21 +165,22 @@ pub struct BountyTriggerApplicationReviewCommand {
     pub app_id: u64,
 }
 
-#[async_trait]
-impl<T: Runtime + Org + Vote + Donate + Bank + Bounty, P: Pair> Command<T, P>
-    for BountyTriggerApplicationReviewCommand
-where
-    <T as System>::AccountId: Ss58Codec,
-    <T as Vote>::VoteId: Display,
-    <T as Bounty>::BountyId: From<u64> + Display,
-{
-    async fn exec(&self, client: &dyn AbstractClient<T, P>) -> Result<()> {
+impl BountyTriggerApplicationReviewCommand {
+    pub async fn exec<R: Runtime + Bounty, C: BountyClient<R>>(
+        &self,
+        client: &C,
+    ) -> Result<(), C::Error>
+    where
+        <R as System>::AccountId: Ss58Codec,
+        <R as Bounty>::BountyId: From<u64> + Display,
+    {
         let event = client
             .account_triggers_application_review(
                 self.bounty_id.into(),
                 self.app_id.into(),
             )
-            .await?;
+            .await
+            .map_err(Error::Client)?;
         println!(
             "AccountId {:?} triggered review for bounty {} application {} with application state {:?}",
             event.trigger, event.bounty_id, event.application_id, event.application_state
@@ -195,21 +195,23 @@ pub struct BountySudoApproveApplicationCommand {
     pub app_id: u64,
 }
 
-#[async_trait]
-impl<T: Runtime + Org + Vote + Donate + Bank + Bounty, P: Pair> Command<T, P>
-    for BountySudoApproveApplicationCommand
-where
-    <T as System>::AccountId: Ss58Codec,
-    <T as Vote>::VoteId: Display,
-    <T as Bounty>::BountyId: From<u64> + Display,
-{
-    async fn exec(&self, client: &dyn AbstractClient<T, P>) -> Result<()> {
+impl BountySudoApproveApplicationCommand {
+    pub async fn exec<R: Runtime + Bounty, C: BountyClient<R>>(
+        &self,
+        client: &C,
+    ) -> Result<(), C::Error>
+    where
+        <R as System>::AccountId: Ss58Codec,
+        <R as Vote>::VoteId: Display,
+        <R as Bounty>::BountyId: From<u64> + Display,
+    {
         let event = client
             .account_sudo_approves_application(
                 self.bounty_id.into(),
                 self.app_id.into(),
             )
-            .await?;
+            .await
+            .map_err(Error::Client)?;
         println!(
             "AccountId {:?} sudo approved bounty {} application {} with application state {:?}",
             event.sudo, event.bounty_id, event.application_id, event.application_state
@@ -224,18 +226,20 @@ pub struct BountyPollApplicationCommand {
     pub app_id: u64,
 }
 
-#[async_trait]
-impl<T: Runtime + Org + Vote + Donate + Bank + Bounty, P: Pair> Command<T, P>
-    for BountyPollApplicationCommand
-where
-    <T as System>::AccountId: Ss58Codec,
-    <T as Vote>::VoteId: Display,
-    <T as Bounty>::BountyId: From<u64> + Display,
-{
-    async fn exec(&self, client: &dyn AbstractClient<T, P>) -> Result<()> {
+impl BountyPollApplicationCommand {
+    pub async fn exec<R: Runtime + Bounty, C: BountyClient<R>>(
+        &self,
+        client: &C,
+    ) -> Result<(), C::Error>
+    where
+        <R as System>::AccountId: Ss58Codec,
+        <R as Vote>::VoteId: Display,
+        <R as Bounty>::BountyId: From<u64> + Display,
+    {
         let event = client
             .poll_application(self.bounty_id.into(), self.app_id.into())
-            .await?;
+            .await
+            .map_err(Error::Client)?;
         println!(
             "AccountId {:?} polled bounty {} application {} with application state {:?}",
             event.poller, event.bounty_id, event.application_id, event.application_state
@@ -248,34 +252,34 @@ where
 pub struct BountySubmitMilestoneCommand {
     pub bounty_id: u64,
     pub application_id: u64,
-    pub submission_reference: String,
+    // submission reference
+    pub repo_owner: String,
+    pub repo_name: String,
+    pub issue_number: u64,
     pub amount_requested: u128,
 }
 
-#[async_trait]
-impl<T: Runtime + Org + Vote + Donate + Bank + Bounty, P: Pair> Command<T, P>
-    for BountySubmitMilestoneCommand
-where
-    <T as System>::AccountId: Ss58Codec,
-    <T as Org>::IpfsReference: From<CidBytes> + Debug,
-    <T as Bank>::Currency: From<u128> + Display,
-    <T as Bounty>::BountyId: From<u64> + Display,
-{
-    async fn exec(&self, client: &dyn AbstractClient<T, P>) -> Result<()> {
-        let submission_reference: CidBytes = {
-            let content = self.submission_reference.as_bytes();
-            let hash = Blake2b256::digest(&content[..]);
-            let cid = Cid::new_v1(Codec::Raw, hash);
-            CidBytes::from(&cid)
-        };
+impl BountySubmitMilestoneCommand {
+    pub async fn exec<R: Runtime + Bounty, C: BountyClient<R>>(
+        &self,
+        client: &C,
+    ) -> Result<(), C::Error>
+    where
+        <R as System>::AccountId: Ss58Codec,
+        <R as Bank>::Currency: From<u128> + Display,
+        <R as Bounty>::BountyId: From<u64> + Display,
+    {
         let event = client
             .submit_milestone(
                 self.bounty_id.into(),
                 self.application_id.into(),
-                submission_reference.into(),
+                (*self.repo_owner).to_string(),
+                (*self.repo_name).to_string(),
+                self.issue_number,
                 self.amount_requested.into(),
             )
-            .await?;
+            .await
+            .map_err(Error::Client)?;
         println!(
             "AccountId {:?} submitted a milestone for bounty {} application {} milestone {} for amount {}",
             event.submitter, event.bounty_id, event.application_id, event.new_milestone_id, event.amount_requested,
@@ -290,19 +294,21 @@ pub struct BountyTriggerMilestoneReviewCommand {
     pub milestone_id: u64,
 }
 
-#[async_trait]
-impl<T: Runtime + Org + Vote + Donate + Bank + Bounty, P: Pair> Command<T, P>
-    for BountyTriggerMilestoneReviewCommand
-where
-    <T as Bounty>::BountyId: From<u64> + Display,
-{
-    async fn exec(&self, client: &dyn AbstractClient<T, P>) -> Result<()> {
+impl BountyTriggerMilestoneReviewCommand {
+    pub async fn exec<R: Runtime + Bounty, C: BountyClient<R>>(
+        &self,
+        client: &C,
+    ) -> Result<(), C::Error>
+    where
+        <R as Bounty>::BountyId: From<u64> + Display,
+    {
         let event = client
             .trigger_milestone_review(
                 self.bounty_id.into(),
                 self.milestone_id.into(),
             )
-            .await?;
+            .await
+            .map_err(Error::Client)?;
         println!(
             "AccountId {:?} triggered a milestone review for bounty {} milestone {} with state {:?}",
             event.trigger, event.bounty_id, event.milestone_id, event.milestone_state,
@@ -317,19 +323,21 @@ pub struct BountySudoApproveMilestoneCommand {
     pub milestone_id: u64,
 }
 
-#[async_trait]
-impl<T: Runtime + Org + Vote + Donate + Bank + Bounty, P: Pair> Command<T, P>
-    for BountySudoApproveMilestoneCommand
-where
-    <T as Bounty>::BountyId: From<u64> + Display,
-{
-    async fn exec(&self, client: &dyn AbstractClient<T, P>) -> Result<()> {
+impl BountySudoApproveMilestoneCommand {
+    pub async fn exec<R: Runtime + Bounty, C: BountyClient<R>>(
+        &self,
+        client: &C,
+    ) -> Result<(), C::Error>
+    where
+        <R as Bounty>::BountyId: From<u64> + Display,
+    {
         let event = client
             .sudo_approves_milestone(
                 self.bounty_id.into(),
                 self.milestone_id.into(),
             )
-            .await?;
+            .await
+            .map_err(Error::Client)?;
         println!(
             "AccountId {:?} sudo approved bounty {} milestone {} with state {:?}",
             event.sudo, event.bounty_id, event.milestone_id, event.milestone_state,
@@ -344,16 +352,18 @@ pub struct BountyPollMilestoneCommand {
     pub milestone_id: u64,
 }
 
-#[async_trait]
-impl<T: Runtime + Org + Vote + Donate + Bank + Bounty, P: Pair> Command<T, P>
-    for BountyPollMilestoneCommand
-where
-    <T as Bounty>::BountyId: From<u64> + Display,
-{
-    async fn exec(&self, client: &dyn AbstractClient<T, P>) -> Result<()> {
+impl BountyPollMilestoneCommand {
+    pub async fn exec<R: Runtime + Bounty, C: BountyClient<R>>(
+        &self,
+        client: &C,
+    ) -> Result<(), C::Error>
+    where
+        <R as Bounty>::BountyId: From<u64> + Display,
+    {
         let event = client
             .poll_milestone(self.bounty_id.into(), self.milestone_id.into())
-            .await?;
+            .await
+            .map_err(Error::Client)?;
         println!(
             "AccountId {:?} polled bounty {} milestone {} with state {:?}",
             event.poller,