+++ client/cli/src/lib.rs
@@ -1,9 +1,9 @@
 pub mod bank;
-// pub mod bounty;
+pub mod bounty;
 pub mod donate;
 mod error;
-// pub mod org;
+pub mod org;
 pub mod shares;
-// pub mod vote;
+pub mod vote;
 
 pub use crate::error::*;