+++ client/cli/src/vote.rs
@@ -1,36 +1,25 @@
-use crate::{
-    async_trait,
-    error::Error,
-    AbstractClient,
-    Bank,
-    Bounty,
-    Command,
-    Donate,
-    Org,
-    Pair,
-    Permill,
+use crate::error::{
+    Error,
     Result,
-    Runtime,
-    Vote,
 };
 use clap::Clap;
 use core::fmt::{
     Debug,
     Display,
 };
-use libipld::{
-    cid::{
-        Cid,
-        Codec,
-    },
-    multihash::Blake2b256,
-};
 use substrate_subxt::{
     sp_core::crypto::Ss58Codec,
     system::System,
+    Runtime,
 };
-use util::vote::VoterView;
-use utils_identity::cid::CidBytes;
+use sunshine_bounty_client::{
+    org::Org,
+    vote::{
+        Vote,
+        VoteClient,
+    },
+};
+use sunshine_bounty_utils::vote::VoterView;
 
 #[derive(Clone, Debug, Clap)]
 pub struct VoteCreateSignalThresholdCommand {
@@ -41,34 +30,25 @@ pub struct VoteCreateSignalThresholdCommand {
     pub duration: Option<u32>,
 }
 
-#[async_trait]
-impl<T: Runtime + Org + Vote + Donate + Bank + Bounty, P: Pair> Command<T, P>
-    for VoteCreateSignalThresholdCommand
-where
-    <T as System>::AccountId: Ss58Codec,
-    <T as System>::BlockNumber: From<u32> + Display,
-    <T as Org>::OrgId: From<u64> + Display,
-    <T as Org>::IpfsReference: From<CidBytes> + Debug,
-    <T as Vote>::Signal: From<u64> + Display,
-    <T as Vote>::VoteId: From<u64> + Display,
-{
-    async fn exec(&self, client: &dyn AbstractClient<T, P>) -> Result<()> {
-        let topic: Option<T::IpfsReference> =
-            if let Some(topic_ref) = &self.topic {
-                let content = topic_ref.as_bytes();
-                let hash = Blake2b256::digest(&content[..]);
-                let cid = Cid::new_v1(Codec::Raw, hash);
-                Some(CidBytes::from(&cid).into())
-            } else {
-                None
-            };
-        let turnout_requirement: Option<T::Signal> =
+impl VoteCreateSignalThresholdCommand {
+    pub async fn exec<R: Runtime + Vote, C: VoteClient<R>>(
+        &self,
+        client: &C,
+    ) -> Result<(), C::Error>
+    where
+        <R as System>::AccountId: Ss58Codec,
+        <R as System>::BlockNumber: From<u32>,
+        <R as Org>::OrgId: From<u64> + Display,
+        <R as Vote>::Signal: From<u64> + Display,
+        <R as Vote>::VoteId: Display,
+    {
+        let turnout_requirement: Option<R::Signal> =
             if let Some(req) = self.turnout_requirement {
                 Some(req.into())
             } else {
                 None
             };
-        let duration: Option<<T as System>::BlockNumber> =
+        let duration: Option<<R as System>::BlockNumber> =
             if let Some(req) = self.duration {
                 Some(req.into())
             } else {
@@ -76,13 +56,14 @@ where
             };
         let event = client
             .create_signal_threshold_vote(
-                topic,
+                self.topic.clone(),
                 self.organization.into(),
                 self.support_requirement.into(),
                 turnout_requirement,
                 duration,
             )
-            .await?;
+            .await
+            .map_err(Error::Client)?;
         println!(
             "Account {} created a signal threshold vote for OrgId {} with VoteId {}",
             event.caller, event.org, event.new_vote_id
@@ -100,43 +81,18 @@ pub struct VoteCreatePercentThresholdCommand {
     pub duration: Option<u32>,
 }
 
-fn u8_to_permill(u: u8) -> Result<Permill> {
-    if u > 0u8 && u < 100u8 {
-        Ok(Permill::from_percent(u.into()))
-    } else {
-        Err(Error::VotePercentThresholdInputBoundError)
-    }
-}
-
-#[async_trait]
-impl<T: Runtime + Org + Vote + Donate + Bank + Bounty, P: Pair> Command<T, P>
-    for VoteCreatePercentThresholdCommand
-where
-    <T as System>::AccountId: Ss58Codec,
-    <T as System>::BlockNumber: From<u32> + Display,
-    <T as Org>::OrgId: From<u64> + Display,
-    <T as Org>::IpfsReference: From<CidBytes> + Debug,
-    <T as Vote>::VoteId: From<u64> + Display,
-{
-    async fn exec(&self, client: &dyn AbstractClient<T, P>) -> Result<()> {
-        let topic: Option<T::IpfsReference> =
-            if let Some(topic_ref) = &self.topic {
-                let content = topic_ref.as_bytes();
-                let hash = Blake2b256::digest(&content[..]);
-                let cid = Cid::new_v1(Codec::Raw, hash);
-                Some(CidBytes::from(&cid).into())
-            } else {
-                None
-            };
-        let support_threshold = u8_to_permill(self.support_threshold)?;
-        let turnout_threshold: Option<Permill> =
-            if let Some(req) = self.turnout_threshold {
-                let ret = u8_to_permill(req)?;
-                Some(ret)
-            } else {
-                None
-            };
-        let duration: Option<<T as System>::BlockNumber> =
+impl VoteCreatePercentThresholdCommand {
+    pub async fn exec<R: Runtime + Vote, C: VoteClient<R>>(
+        &self,
+        client: &C,
+    ) -> Result<(), C::Error>
+    where
+        <R as System>::AccountId: Ss58Codec,
+        <R as System>::BlockNumber: From<u32>,
+        <R as Org>::OrgId: From<u64> + Display,
+        <R as Vote>::VoteId: Display,
+    {
+        let duration: Option<<R as System>::BlockNumber> =
             if let Some(req) = self.duration {
                 Some(req.into())
             } else {
@@ -144,13 +100,14 @@ where
             };
         let event = client
             .create_percent_threshold_vote(
-                topic,
+                self.topic.clone(),
                 self.organization.into(),
-                support_threshold,
-                turnout_threshold,
+                self.support_threshold,
+                self.turnout_threshold,
                 duration,
             )
-            .await?;
+            .await
+            .map_err(Error::Client)?;
         println!(
             "Account {} created a percent threshold vote for OrgId {} with VoteId {}",
             event.caller, event.org, event.new_vote_id
@@ -166,28 +123,17 @@ pub struct VoteCreateUnanimousConsentCommand {
     pub duration: Option<u32>,
 }
 
-#[async_trait]
-impl<T: Runtime + Org + Vote + Donate + Bank + Bounty, P: Pair> Command<T, P>
-    for VoteCreateUnanimousConsentCommand
-where
-    <T as System>::AccountId: Ss58Codec,
-    <T as System>::BlockNumber: From<u32> + Display,
-    <T as Org>::OrgId: From<u64> + Display,
-    <T as Org>::IpfsReference: From<CidBytes> + Debug,
-    <T as Vote>::Signal: From<u64> + Display,
-    <T as Vote>::VoteId: From<u64> + Display,
-{
-    async fn exec(&self, client: &dyn AbstractClient<T, P>) -> Result<()> {
-        let topic: Option<T::IpfsReference> =
-            if let Some(topic_ref) = &self.topic {
-                let content = topic_ref.as_bytes();
-                let hash = Blake2b256::digest(&content[..]);
-                let cid = Cid::new_v1(Codec::Raw, hash);
-                Some(CidBytes::from(&cid).into())
-            } else {
-                None
-            };
-        let duration: Option<<T as System>::BlockNumber> =
+impl VoteCreateUnanimousConsentCommand {
+    pub async fn exec<R: Runtime + Vote, C: VoteClient<R>>(
+        &self,
+        client: &C,
+    ) -> Result<(), C::Error>
+    where
+        <R as System>::AccountId: Ss58Codec,
+        <R as Org>::OrgId: From<u64> + Display,
+        <R as Vote>::VoteId: Display,
+    {
+        let duration: Option<<R as System>::BlockNumber> =
             if let Some(req) = self.duration {
                 Some(req.into())
             } else {
@@ -195,11 +141,12 @@ where
             };
         let event = client
             .create_unanimous_consent_vote(
-                topic,
+                self.topic.clone(),
                 self.organization.into(),
                 duration,
             )
-            .await?;
+            .await
+            .map_err(Error::Client)?;
         println!(
             "Account {} created a unanimous consent vote for OrgId {} with VoteId {}",
             event.caller, event.org, event.new_vote_id
@@ -215,32 +162,29 @@ pub struct VoteSubmitCommand {
     pub justification: Option<String>,
 }
 
-#[async_trait]
-impl<T: Runtime + Org + Vote + Donate + Bank + Bounty, P: Pair> Command<T, P>
-    for VoteSubmitCommand
-where
-    <T as System>::AccountId: Ss58Codec,
-    <T as Org>::IpfsReference: From<CidBytes> + Debug,
-    <T as Vote>::VoteId: From<u64> + Display,
-{
-    async fn exec(&self, client: &dyn AbstractClient<T, P>) -> Result<()> {
-        let justification: Option<T::IpfsReference> =
-            if let Some(topic_ref) = &self.justification {
-                let content = topic_ref.as_bytes();
-                let hash = Blake2b256::digest(&content[..]);
-                let cid = Cid::new_v1(Codec::Raw, hash);
-                Some(CidBytes::from(&cid).into())
-            } else {
-                None
-            };
+impl VoteSubmitCommand {
+    pub async fn exec<R: Runtime + Vote, C: VoteClient<R>>(
+        &self,
+        client: &C,
+    ) -> Result<(), C::Error>
+    where
+        <R as System>::AccountId: Ss58Codec,
+        <R as Org>::OrgId: From<u64> + Display,
+        <R as Vote>::VoteId: From<u64> + Display,
+    {
         let voter_view = match self.direction {
             0u8 => VoterView::Against, // 0 == false
             1u8 => VoterView::InFavor, // 1 == true
             _ => VoterView::Abstain,
         };
         let event = client
-            .submit_vote(self.vote_id.into(), voter_view, justification.into())
-            .await?;
+            .submit_vote(
+                self.vote_id.into(),
+                voter_view,
+                self.justification.clone(),
+            )
+            .await
+            .map_err(Error::Client)?;
         println!(
             "Account {} voted with view {:?} in VoteId {}",
             event.voter, event.view, event.vote_id