+++ client/client/src/error.rs
@@ -6,4 +6,6 @@ pub enum Error {
     EventNotFound,
     #[error("Number cannot be parsed from string")]
     ParseIntError,
+    #[error("Vote percent threshold input must be 0 < x < 100")]
+    VotePercentThresholdInputBoundError,
 }