+++ CHANGELOG.md
@@ -6,15 +6,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- minimal `sunshine-treasury` pallet, extracted from `sunshine-donate` pallet
+- `OffchainClient` to use ipfs-embed in a structured way (`client/client/src/lib.rs`)
+
 ## [0.1.1] - 2020-07-15
 
-- released everything except `client/*` and the `bin/{cli, client}`
 - fixed lossy donate module by adding explicit remainder return value
 - added percentage thresholds to the vote module which rounds up (instead of down or to the nearest value) (see tests)
 
 ## [0.0.6] - 2020-07-07
 
-- runtime and node in `bin/runtime` and `bin/node`
 - generic client library in `client/client` with example usage in `bin/client`
 - generic cli library in `client/cli` with example usage in `bin/cli`